@@ -1,4 +1,7 @@
-﻿using SaberFactory.Helpers;
+﻿using System;
+using IPA.Utilities;
+using SaberFactory.Helpers;
+using SiraUtil;
 using UnityEngine;
 
 namespace SaberFactory.Instances.Trail
@@ -13,27 +16,68 @@ namespace SaberFactory.Instances.Trail
         protected SaberTrailRenderer _trailRenderer;
         protected InstanceTrailData _instanceTrailData;
 
+        private readonly SaberTrail _backupTrail;
+
         public TrailHandler(GameObject gameobject)
         {
             TrailInstance = gameobject.AddComponent<SFTrail>();
         }
 
+        public TrailHandler(GameObject gameobject, SaberTrail backupTrail) : this(gameobject)
+        {
+            _backupTrail = backupTrail;
+        }
+
         public void CreateTrail()
         {
-            var trailInitData = new SFTrail.TrailInitData
+            if (_trailRenderer is null)
             {
-                TrailColor = Color.white,
-                TrailLength = _instanceTrailData.Length,
-                TrailPrefab = _trailRenderer,
-                Whitestep = _instanceTrailData.WhiteStep
-            };
+                throw new ArgumentNullException(nameof(_trailRenderer), "Trail Renderer is not specified");
+            }
 
-            TrailInstance.Setup(
-                trailInitData,
-                _instanceTrailData.Material.Material,
-                _instanceTrailData.PointStart,
-                _instanceTrailData.PointEnd
+            if (_instanceTrailData is null)
+            {
+
+                if (_backupTrail is null) return;
+
+                var trailStart = TrailInstance.gameObject.CreateGameObject("Trail Start");
+                var trailEnd = TrailInstance.gameObject.CreateGameObject("TrailEnd");
+                trailEnd.transform.localPosition = new Vector3(0, 0, 1);
+
+                var material = _trailRenderer.GetField<MeshRenderer, SaberTrailRenderer>("_meshRenderer").material;
+
+                var trailInitData = new SFTrail.TrailInitData
+                {
+                    TrailColor = Color.white,
+                    TrailLength = 15,
+                    TrailPrefab = _trailRenderer,
+                    Whitestep = 0.02f
+                };
+
+                TrailInstance.Setup(
+                    trailInitData,
+                    material,
+                    trailStart.transform,
+                    trailEnd.transform
                 );
+            }
+            else
+            {
+                var trailInitData = new SFTrail.TrailInitData
+                {
+                    TrailColor = Color.white,
+                    TrailLength = _instanceTrailData.Length,
+                    TrailPrefab = _trailRenderer,
+                    Whitestep = _instanceTrailData.WhiteStep
+                };
+
+                TrailInstance.Setup(
+                    trailInitData,
+                    _instanceTrailData.Material.Material,
+                    _instanceTrailData.PointStart,
+                    _instanceTrailData.PointEnd
+                );
+            }
         }
 
         public void DestroyTrail()
@@ -53,8 +97,14 @@ namespace SaberFactory.Instances.Trail
 
         public void SetColor(Color color)
         {
-            if (TrailInstance == null) return;
-            TrailInstance.Color = color;
+            if (TrailInstance is SFTrail sfTrail)
+            {
+                sfTrail.Color = color;
+            }
+            else
+            {
+                TrailInstance?.SetField("_color", color);
+            }
         }
     }
 }
